requirements.txt
-e backend
alembic
gocept.cache
gocept.testdb
passlib
psycopg2
pyramid
pyramid_tm
sqlalchemy
risclog.sqlalchemy
transaction
waitress
zope.sqlalchemy
